Update model references and integrate Coinpay service
This update changes the naming method of models in various parts of the system. Model names have been simplified for clarity. Additionally, this integrates a CoinPay service to handle the sending of funds in the wallet-removal component. The email column in the network component has been commented out as part of this commit.